so_crawl.crawl now checks validity before adding snippet.
